requirements.txt
dotenv
Django>=4.2.0
django-ckeditor>=6.7.0
djangorestframework>=3.14.0
googletrans>=3.1.0a0
django-redis>=5.4.0
redis>=5.0.1
pytest>=7.4.0
pytest-django>=4.5.2
flake8>=6.1.0